cast to float added

+++ core/src/ArcadeCore/Layout.cpp
@@ -8,7 +8,7 @@
 #include "core/include/ArcadeCore/Layout.hpp"
 
 Layout::Layout(DisplayLibrary *l):
-_backgroundBody({0, (l->_window->height() - l->toUnit(70)), l->_window->width(), l->toUnit(70)})
+_backgroundBody({0, ((float)(l->_window->height() - l->toUnit(70))), (float)(l->_window->width()), (float)(l->toUnit(70))})
 {
 }
